requirements.txt
amqp==2.5.2
asgiref==3.2.4
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.1
decorator==4.4.2
Django==3.0.4
django-redis==4.11.0
djangorestframework==3.11.0
importlib-metadata==1.5.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
kafka-python==2.0.1
kombu==4.6.8
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2019.3
redis==3.4.1
schedule==0.6.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
vine==1.3.0
wcwidth==0.1.8
websocket-client==0.57.0
zipp==3.1.0
